VIDSOL-575 settings module - stats viewer and send sender stats track toggle
